pyproject.toml
[project]
name = "repo-sustain-lab"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-cloud-bigquery>=3.40.0",
    "db-dtypes>=1.0.0",
    "pandas>=3.0.0",
    "pyarrow>=23.0.0",
    "python-dotenv>=1.2.1",
]
